Started work on cleaning up comments.

+++ src/PDFiumCoreBindingsGenerator/Program.cs
@@ -16,6 +16,7 @@ namespace PDFiumCoreBindingsGenerator
     {
         private static WebClient _client;
 
+        private static bool download = false;
         private class LibInfo
         {
             public string PackageName { get; }
@@ -96,8 +97,15 @@ namespace PDFiumCoreBindingsGenerator
 
             Console.WriteLine("Complete.");
 
-            if(Directory.Exists(destinationLibraryPath))
-                Directory.Delete(destinationLibraryPath, true);
+            if (Directory.Exists(destinationLibraryPath))
+            {
+                if(download)
+                    Directory.Delete(destinationLibraryPath, true);
+            }
+            else
+            {
+                download = true;
+            }
 
             Directory.CreateDirectory(destinationLibraryPath);
 
@@ -215,6 +223,9 @@ namespace PDFiumCoreBindingsGenerator
             var filename = Path.GetFileName(uri.LocalPath);
             var fullFilePath = Path.Combine(baseDestination, filename);
             var destinationDirPath = Path.Combine(baseDestination, Path.GetFileNameWithoutExtension(filename));
+
+            if (!download)
+                return destinationDirPath;
 
             if (File.Exists(fullFilePath))
                 File.Delete(fullFilePath);
